Modify entity hierarchy
