+++ src/main/java/kimjinung/commerce/controller/api/MemberApiController.java
@@ -1,6 +1,6 @@
 package kimjinung.commerce.controller.api;
 
-import kimjinung.commerce.controller.api.error.ErrorResult;
+import kimjinung.commerce.dto.error.ErrorResult;
 import kimjinung.commerce.dto.member.*;
 import kimjinung.commerce.exception.InvalidRequestException;
 import kimjinung.commerce.exception.MemberJoinInvalidArgumentException;
@@ -11,7 +11,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
-import org.springframework.validation.FieldError;
 import org.springframework.validation.ObjectError;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -22,37 +21,11 @@ import java.util.List;
 
 @Slf4j
 @RequiredArgsConstructor
-@RequestMapping("/member")
+@RequestMapping("/api/member")
 @RestController
 public class MemberApiController {
-    /*
-    오브젝트 에러: 요청 실패
-    필드 에러: 검증 실패
-    를 구분해서 예외 처리를 한다.
 
-     */
     private final MemberService memberService;
-
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(MemberNotFoundException.class)
-    public ErrorResult memberNotFoundException(MemberNotFoundException e) {
-        log.info("[Exception] ", e);
-        return new ErrorResult("NOT_FOUND", e.getMessage());
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(InvalidRequestException.class)
-    public ErrorResult invalidRequestException(InvalidRequestException e) {
-        log.info("[Exception ", e);
-        return new ErrorResult("INVALID_ARGUMENT", e.getMessage());
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(RuntimeException.class)
-    public ErrorResult runtimeException(RuntimeException e) {
-        log.info("[Exception] ", e);
-        return new ErrorResult("BAD_REQUEST", e.getMessage());
-    }
 
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
     @ExceptionHandler(Exception.class)
