--- src/main/java/kimjinung/commerce/controller/web/OrderController.java
@@ -1,4 +1,0 @@
-package kimjinung.commerce.controller.web;
-
-public class OrderController {
-}
