--- src/main/java/kimjinung/commerce/controller/web/ProductController.java
@@ -1,4 +1,0 @@
-package kimjinung.commerce.controller.web;
-
-public class ProductController {
-}
