+++ src/main/java/kimjinung/commerce/domain/Member.java
@@ -1,5 +1,6 @@
 package kimjinung.commerce.domain;
 
+import kimjinung.commerce.domain.common.BaseEntity;
 import lombok.Getter;
 import org.hibernate.annotations.GenericGenerator;
 
@@ -8,7 +9,7 @@ import java.util.UUID;
 
 @Getter
 @Entity
-public class Member extends BaseEntity{
+public class Member extends BaseEntity {
 
     @Id
     @GeneratedValue(generator = "uuid")
